[ci] Shellcheck fixes for make-distrib.sh

+++ dev/make-distrib.sh
@@ -72,7 +72,7 @@ do
             ;;
         "--codename")
             shift
-            VERSION_CODENAME=$1
+            VERSION_CODENAME="$1"
             ;;
         "--ci-link")
             CI_LINK=yes
@@ -89,7 +89,7 @@ done
 # Target Names
 
 VERSION=$(cat VERSION)
-VERSION_SAFE=${VERSION/~/-}
+VERSION_SAFE="${VERSION/~/-}"
 
 FRAMAC="frama-c-$VERSION_SAFE-$VERSION_CODENAME"
 if [ "$USE_STASH" == "yes" ]; then
@@ -114,7 +114,7 @@ fi
 IGNORED_FILES="$(git ls-files --ignored --exclude-standard -c)"
 if [ "" != "$IGNORED_FILES" ]; then
   echo "Some versioned files are ignored by .gitignore:"
-  echo $IGNORED_FILES
+  echo "$IGNORED_FILES"
   exit 2
 fi
 
@@ -159,9 +159,13 @@ echo "----------------------------------------------------------------"
 ################################################################################
 # Warn if there are uncommitted changes (will not be taken into account)
 
+# We do want word splitting, disable SC2046
+# shellcheck disable=SC2046
 GIT_STATUS="$(git status --porcelain -- $(printf ":!%s\n" "${EXTERNAL_PLUGINS[@]}"))"
-if [ "" != "$GIT_STATUS" -a "$USE_STASH" != "yes" ]; then
+if [ "" != "$GIT_STATUS" ] && [ "$USE_STASH" != "yes" ]; then
   echo "WARNING: uncommitted changes will be IGNORED when making archive:"
+  # Unable to replace sed with ${var//search/replace} here, disable SC2001
+  # shellcheck disable=SC2001
   echo "$GIT_STATUS" | sed 's/^/  /'
   echo "----------------------------------------------------------------"
 fi
@@ -174,7 +178,7 @@ if [ "$USE_STASH" == "yes" ]; then
     ARCHIVE_COMMIT=$(git stash create)
 fi
 
-git archive ${ARCHIVE_COMMIT:-HEAD} -o $FRAMAC_TAR --prefix "$FRAMAC/"
+git archive "${ARCHIVE_COMMIT:-HEAD}" -o "$FRAMAC_TAR" --prefix "$FRAMAC/"
 
 ################################################################################
 # Add external plugins to archive
@@ -187,10 +191,10 @@ fi
 for plugin in "${EXTERNAL_PLUGINS[@]}"
 do
     echo "  $plugin"
-    PLUGIN_TAR="$(basename $plugin).tar"
-    git -C $plugin archive HEAD -o $PLUGIN_TAR --prefix "$FRAMAC/$plugin/"
-    $TAR --concatenate --file=$FRAMAC_TAR "$plugin/$PLUGIN_TAR"
-    rm -rf "$plugin/$PLUGIN_TAR"
+    PLUGIN_TAR="$(basename "$plugin").tar"
+    git -C "$plugin" archive HEAD -o "$PLUGIN_TAR" --prefix "$FRAMAC/$plugin/"
+    $TAR --concatenate --file="$FRAMAC_TAR" "$plugin/$PLUGIN_TAR"
+    rm -rf "${plugin:?}/${PLUGIN_TAR:?}"
 done
 
 if [ "${#EXTERNAL_PLUGINS[@]}" -gt 0 ]
@@ -209,14 +213,14 @@ HEADER_SPEC="header-spec.txt"
 git ls-files |\
 git check-attr --stdin export-ignore |\
 grep -v "export-ignore: set" | awk -F ': ' '{print $1}' |\
-git check-attr --stdin header_spec > $HEADER_SPEC
+git check-attr --stdin header_spec > "$HEADER_SPEC"
 
 for plugin in "${EXTERNAL_PLUGINS[@]}" ; do
-  git -C $plugin ls-files |\
-  git -C $plugin check-attr --stdin export-ignore |\
+  git -C "$plugin" ls-files |\
+  git -C "$plugin" check-attr --stdin export-ignore |\
   grep -v "export-ignore: set" | awk -F ': ' '{print $1}' |\
-  git -C $plugin check-attr --stdin header_spec |\
-  xargs -n3 printf "$plugin/%s %s %s\n" >> $HEADER_SPEC
+  git -C "$plugin" check-attr --stdin header_spec |\
+  xargs -n3 printf "$plugin/%s %s %s\n" >> "$HEADER_SPEC"
 done
 
 ################################################################################
@@ -238,9 +242,11 @@ done
 echo "Check headers..."
 
 TMP_DIR=$(mktemp -d)
-$TAR xf $FRAMAC_TAR -C $TMP_DIR
+$TAR xf "$FRAMAC_TAR" -C "$TMP_DIR"
 
 # Check
+# We do want globbing and word splitting, disable SC2086
+# shellcheck disable=SC2086
 $HDRCK $CHECK_HEADER_OPT -spec-format="3-fields-by-line" -C "$TMP_DIR/$FRAMAC" $HEADER_SPEC
 
 ################################################################################
@@ -248,25 +254,25 @@ $HDRCK $CHECK_HEADER_OPT -spec-format="3-fields-by-line" -C "$TMP_DIR/$FRAMAC" $
 
 echo "Finalizing archive..."
 
-echo $VERSION_SAFE > $TMP_DIR/$FRAMAC/VERSION
-echo $VERSION_CODENAME > $TMP_DIR/$FRAMAC/VERSION_CODENAME
+echo "$VERSION_SAFE" > "$TMP_DIR/$FRAMAC/VERSION"
+echo "$VERSION_CODENAME" > "$TMP_DIR/$FRAMAC/VERSION_CODENAME"
 
 DATE="$(date +%F)"
 
-$TAR czf $FRAMAC_TAR.gz -C $TMP_DIR $FRAMAC \
+$TAR czf "$FRAMAC_TAR.gz" -C "$TMP_DIR" "$FRAMAC" \
   --numeric-owner --owner=0 --group=0 --sort=name --mode='a+rw' \
   --mtime="$DATE Z"
 
 if [[ "$CI_LINK" == "yes" ]]; then
-  ln $FRAMAC_TAR.gz "frama-c.tar.gz"
+  ln "$FRAMAC_TAR.gz" "frama-c.tar.gz"
 fi
 
 ################################################################################
 # Cleaning
 
-rm -rf $HEADER_SPEC
-rm -rf $FRAMAC_TAR
-rm -rf $TMP_DIR
+rm -rf "$HEADER_SPEC"
+rm -rf "$FRAMAC_TAR"
+rm -rf "$TMP_DIR"
 
 echo "Generated: $FRAMAC_TAR.gz"
 echo "----------------------------------------------------------------"
